requirements.txt
# Core Django and utilities
Django==4.2
python-dotenv==1.0.0

# Video/Audio Processing
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
torchvision==0.18.1+cpu
torchaudio==2.5.1+cpu
openai-whisper==20231117
ffmpeg-python==0.2.0

# OpenAI
openai==1.12.0

# Other dependencies
Pillow==10.2.0  # For image handling
django-crispy-forms==2.1  # If you're using crispy forms 